Move inline completion prompt into txt file
