+++ codegpt-core/src/main/resources/prompts/inline-completion-prompt.txt
@@ -1,0 +1,1 @@
+{pre}{codeBefore}{suf}{codeAfter}{mid}
